+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - `isValidSwissSocialInsuranceNumber` is now named properly
+- `isValidSwissIbanNumber` now also allows IBAN numbers with letters
 
 ## [2.1.0] - 2025-09-03